requirements.txt
ansible==2.7.10
asn1crypto==0.24.0
awscli==1.16.146
bcrypt==3.1.6
boto3==1.9.136
botocore==1.12.136
cffi==1.12.3
cloudpickle==0.8.1
colorama==0.3.9
cryptography==2.6.1
docutils==0.14
doit==0.31.1
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
paramiko==2.4.2
pex==1.6.6
pyasn1==0.4.5
pycparser==2.19
pyinotify==0.9.6
PyNaCl==1.3.0
python-dateutil==2.8.0
PyYAML==3.13
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.2
